# src/main/java/com/example/demo/docs/api-spec.yaml
...
openapi: 3.0.3
info:
  title: E-Commerce-API Spec
  description: 電商API功能規劃
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
  /users/login:
    post:
      tags:
        - users
      summary: "使用者登入"
      description: "使用者輸入帳號密碼進行登入"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: "登入成功"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: "#/components/schemas/LoginInfo"
        '400':
          description: "登入失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    email: "帳號或密碼錯誤"
                    password: "帳號或密碼錯誤"
  /users/register:
    post:
      tags:
        - users
      summary: "使用者註冊"
      description: "使用者輸入資料進行註冊"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        '200':
          description: "註冊成功"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: "#/components/schemas/UserInfo"
        '400':
          description: "註冊失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    email: "格式不符合"
                    password: "密碼未輸入"
        '401': 
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
  /users/me:
    get:
      tags:
        - users
      summary: "使用者資料查詢"
      description: "查詢目前使用者資料"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "查詢成功"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: "#/components/schemas/UserInfo"
        '400':
          description: "查詢失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    general: "連線異常請重新登入"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
    put:
      tags:
        - users
      summary: 使用者資料更新
      description: 使用者修改資料或照片
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        '200':
          description: "資料更新成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        '400':
          description: "資料更新失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    image: "照片格是錯誤"
                    birthday: "生日不能大於今日"
                    address: "地址未輸入"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
  /users/me/password:
    patch:
      tags:
        - users
      summary: 使用者修改密碼
      description: 使用者輸入新的密碼進行修改
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        '200':
          description: "密碼修改成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        '400':
          description: "密碼修改失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    oldPassword: "舊密碼錯誤"
                    NewConfirmPassword: "輸入密碼不一致"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
  /users/verify/{token}:
    get:
      tags:
        - users
      summary: "使用者驗證"
      description: "驗證使用者是否登入"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
          description: "JWT Token"
      responses:
        '200':
          description: "驗證成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        '400':
          description: "驗證失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    general: "連線異常請重新登入"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
  /products:
    get:
      tags:
        - products
      summary: "產品列表"
      description: "取得產品列表"
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: categoryId
          required: false
          schema:
            type: integer
          description: "分類ID"
        - in: query
          name: name
          required: false
          schema:
            type: string
          description: "產品名稱"
      responses:
        '200':
          description: "取得產品列表成功"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          products:
                            type: array
                            items:
                              $ref: "#/components/schemas/ProductInfo"
        '400':
          description: "取得產品列表失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    general: "連線異常請重新登入"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
  /products/{productId}:
    get:
      tags:
        - products
      summary: "產品內容"
      description: "取得產品內容"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
          description: "產品ID"
      responses:
        '200':
          description: "取得產品內容成功"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          product:
                            $ref: "#/components/schemas/ProductInfo"
        '400':
          description: "取得產品內容失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    general: "連線異常請重新登入"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
...
  /admin/products/{productId}:
    put:
      tags:
        - admin
      summary: "更新商品"
      description: "管理員更新商品資訊"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
          description: "商品ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInfo"
      responses:
        '200':
          description: "更新商品成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        '400':
          description: "更新商品失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    name: "商品名稱未輸入"
                    price: "價格必須大於0"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
    delete:
      tags:
        - admin
      summary: "刪除商品"
      description: "管理員刪除商品"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
          description: "商品ID"
      responses:
        '200':
          description: "刪除商品成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        '400':
          description: "刪除商品失敗"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                timestamp: "2025-06-23T14:00:00"
                data:
                  errors:
                    general: "連線異常請重新登入"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          $ref: "#/components/responses/forbiddenError"
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        data:
          nullable: true
          type: object
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          example: Test@gmail.com
        password:
          type: string
          example: Aa123456
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: "Test@gmail.com"
        password:
          type: string
          example: Aa123456
    UpdateUserRequest:
      type: object
      properties:
        image:
          type: string
          example: /upload/defaultAvater.jpg
        name:
          type: string
          example: "Test"
        gender:
          type: number
          example: 1
        birthday:
          type: string
          format: date
          example: 2025-06-20
        phone:
          type: string
          example: "0912345678"
        city:
          type: string
          example: "台北市"
        country:
          type: string
          example: "中正區"
        address:
          type: string
          example: "我家巷口"
    UpdatePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
        - newConfirmPassword
      properties:
        oldPassword:
          type: string
          example: Aa123456
        newPassword:
          type: string
          example: Bb123456
        newConfirmPassword:
          type: string
          example: Bb123456
...
    LoginInfo:
      type: object
      properties:
        name:
          type: string
          example: "Test"
        email:
          type: string
          example: "test@gmail.com"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
    UserInfo:
      type: object
      properties:
        email:
          type: string
          example: Test@gmail.com
        name:
          type: string
          example: "Test"
    ProductInfo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "藍芽耳機"
        price:
          type: integer
          example: 2000
        promotionId:
          type: integer
          example: 1
        imageUrl:
          type: string
          example: "/upload/defaultProduct.jpg"
        categoryId:
          type: integer
          example: 1
        inStock:
          type: integer
          example: 20
        rating:
          type: number
          format: double
          example: 4.5
        soldCount:
          type: integer
          example: 36
        shortDescription:
          type: string
          example: "這是最新款的藍芽耳機，目前折價9折熱賣中"
        tags:
          type: array
          example: ["熱賣","新品"]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 
  responses:
    UnauthorizedError:
      description: "未授權"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
          example:
            timestamp: "2025-06-23T14:00:00"
            data: null
    forbiddenError:
      description: "禁止存取"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
          example:
            timestamp: "2025-06-23T14:00:00"
            data: null
...